Update suspense pattern
